Call original swift using full path to avoid conflicts
If `swiftly` is named 'swift' and placed ahead of `swift` in the `$PATH`,
then it ends up being the default script and can, in some cases, end up
in an endless loop when it tries to call the original `swift`
executable. Calling the original swift executable using its full path
eliminates this issue, and also will make it easier to configure the
swift path if necessary.